Fixed DAQ_sync bug
A header byte (0xFF) was added to the data packet, sent from FPGA, which now consists of 4 bytes.
On CyUSB firmware, auto-commit length was increased to 512 bytes (128 frames).
On labview code, VISA read was set at 1024 bytes and a function that discards data packet if header byte is not found, was implemented.